Move fail def integration tests to test_program
